+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 1.48.1
+
+### New features
+
+- Added `\Wizaplace\SDK\Catalog\CatalogService::getProductsByFilters`
+- Added `\Wizaplace\SDK\Catalog\ProductFilter`
+
 ## 1.48.0
 
 ### New features